#16270 move `WebDriverInfo` from `org.openqa.selenium` to `org.openqa.selenium.remote`
Interface `WebDriverInfo` is only needed in `org.openqa.selenium.remote.RemoteWebDriverBuilder`, so it seems reasonable.

Motivation: I added parameter `ClientConfig` to method `WebDriverInfo.createDriver()`, and it doesn't compile anymore because package `org.openqa.selenium` cannot use classes from `org.openqa.selenium.remote.http`.
