+++ java/src/org/openqa/selenium/grid/BUILD.bazel
@@ -137,7 +137,7 @@ java_export(
         "release-artifact",
     ],
     uses = [
-        "org.openqa.selenium.WebDriverInfo",
+        "org.openqa.selenium.remote.WebDriverInfo",
         "org.openqa.selenium.cli.CliCommand",
         "org.openqa.selenium.grid.config.HasRoles",
         "org.openqa.selenium.remote.locators.CustomLocator",
