+++ java/src/org/openqa/selenium/remote/WebDriverInfo.java
@@ -15,9 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium;
+package org.openqa.selenium.remote;
 
 import java.util.Optional;
+import org.openqa.selenium.Beta;
+import org.openqa.selenium.Capabilities;
+import org.openqa.selenium.SessionNotCreatedException;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.http.ClientConfig;
 
 /**
