+++ java/test/org/openqa/selenium/remote/FakeWebDriverInfo.java
@@ -25,7 +25,6 @@ import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.ImmutableCapabilities;
 import org.openqa.selenium.SessionNotCreatedException;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebDriverInfo;
 import org.openqa.selenium.remote.http.ClientConfig;
 
 @AutoService(WebDriverInfo.class)
@@ -85,10 +84,7 @@ public class FakeWebDriverInfo implements WebDriverInfo {
     protected FakeWebDriver() {}
 
     public FakeWebDriver(Capabilities capabilities, ClientConfig clientConfig) {
-      super(
-        command -> null,
-          capabilities,
-          clientConfig);
+      super(command -> null, capabilities, clientConfig);
     }
 
     @Override
